spark-model: split the native-FP8 microtest's BF16 rule under the 500-LoC cap
The tolerance gate added in 7cdb0c172 took
`examples/native_fp8_ffn_w8a8_microtest.rs` to 707 lines, over the repo's
500-LoC-per-file cap. Two required checks read that number: the cap job itself,
and the certification self-test in `security.yml` (reported as 'cargo deny'),
whose 'the 500-LoC cap passes on the real tree' case was the one red case out of
226.

Move the cuBLASLt-vs-kernel acceptance rule — the CUBLAS_* tolerance constants
and the prose that derives them, `to_f64`/`ord`/`bf16_ulp`/
`within_one_bf16_ulp`/`is_sign_flip`, the `Compare` report and `compare`
itself, and the CPU tests that pin all of it — into
`examples/common/native_fp8_bf16_compare.rs`, pulled in by `#[path]` the way
the kda/glm5next microtests already share their harness halves. The example
keeps the GPU half: shapes, weight construction, kernel launches, timing and the
gate verdicts.

That is also the natural seam. The moved half needs no GPU, no kernels and no
model, which is exactly why it is the half with unit tests; the tests now sit
beside the code they pin instead of at the foot of a file whose other 400 lines
they never touch.

Only the four names the example still reads cross the module line as
`pub(crate)` (the three CUBLAS_* constants and `compare`, plus `Compare`'s
fields, which `compare`'s signature makes reachable). The ULP helpers stay
private to the new module.

Behaviour is unchanged: no constant, bound or assertion was edited. 707 -> 413
LoC for the example, 311 for the new module.

Gates: rustfmt clean; the cap scan reproduced locally against the workflow's own
allow-list reports 0 violations.
